feat: Update TTS configuration for improved performance and modify alert response structure

+++ backend/src/config/kokoro.ts
@@ -3,5 +3,5 @@ import { KokoroTTS } from "kokoro-js";
 const model_id = "onnx-community/Kokoro-82M-ONNX";
 export const tts = await KokoroTTS.from_pretrained(model_id, {
     device: "cpu",
-    dtype: "fp32",
+    dtype: "fp16",
 });
